Create a test data seeder as a separate console app
This is an experiment to find a new way to seed initial data.  It's generally not a good idea to do this inside the main application, because you have no control over how many instances of the application may be running and accessing the database at the same time.

Note this is only for seeding sample data that may later be deleted.  If you need to populate the database with static data use a different mechanism like HasData in your entity configurations.